// web/src/api/models/document-preview.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DocumentPreviewChunksInner } from './document-preview-chunks-inner';
// May contain unused imports in some cases
// @ts-ignore
import type { VisionChunk } from './vision-chunk';

/**
 * 
 * @export
 * @interface DocumentPreview
 */
export interface DocumentPreview {
    /**
     * The path to the document object.
     * @type {string}
     * @memberof DocumentPreview
     */
    'doc_object_path'?: string;
    /**
     * The name of the document.
     * @type {string}
     * @memberof DocumentPreview
     */
    'doc_filename'?: string;
    /**
     * The path to the converted PDF object.
     * @type {string}
     * @memberof DocumentPreview
     */
    'converted_pdf_object_path'?: string;
    /**
     * The markdown content of the document.
     * @type {string}
     * @memberof DocumentPreview
     */
    'markdown_content'?: string;
    /**
     * 
     * @type {Array<DocumentPreviewChunksInner>}
     * @memberof DocumentPreview
     */
    'chunks'?: Array<DocumentPreviewChunksInner>;
    /**
     * 
     * @type {Array<VisionChunk>}
     * @memberof DocumentPreview
     */
    'vision_chunks'?: Array<VisionChunk>;
}